Renames the file to get more conventions
